+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/Internal/PrometheusExporter.cs
@@ -66,7 +66,7 @@ internal sealed class PrometheusExporter : BaseExporter<Metric>, IPullMetricExpo
 
     internal int ScrapeResponseCacheDurationMilliseconds { get; }
 
-    internal bool OpenMetricsEnabled { get; }
+    internal bool OpenMetricsRequested { get; set; }
 
     /// <inheritdoc/>
     public override ExportResult Export(in Batch<Metric> metrics)