+++ test/OpenTelemetry.Exporter.Prometheus.HttpListener.Tests/PrometheusHttpListenerTests.cs
@@ -94,17 +94,25 @@ public class PrometheusHttpListenerTests
     [Fact]
     public async Task PrometheusExporterHttpServerIntegration_NoOpenMetrics()
     {
-        await this.RunPrometheusExporterHttpServerIntegrationTest(useOpenMetrics: false);
+        await this.RunPrometheusExporterHttpServerIntegrationTest(acceptHeader: string.Empty);
     }
 
-    private async Task RunPrometheusExporterHttpServerIntegrationTest(bool skipMetrics = false, bool useOpenMetrics = true)
+    [Fact]
+    public async Task PrometheusExporterHttpServerIntegration_UseOpenMetricsVersionHeader()
     {
+        await this.RunPrometheusExporterHttpServerIntegrationTest(acceptHeader: "application/openmetrics-text; version=1.0.0");
+    }
+
+    private async Task RunPrometheusExporterHttpServerIntegrationTest(bool skipMetrics = false, string acceptHeader = "application/openmetrics-text")
+    {
+        var requestOpenMetrics = acceptHeader.StartsWith("application/openmetrics-text");
+
         Random random = new Random();
         int retryAttempts = 5;
         int port = 0;
         string address = null;
 
-        MeterProvider provider;
+        MeterProvider provider = null;
         using var meter = new Meter(this.meterName);
 
         while (retryAttempts-- != 0)
@@ -112,14 +120,24 @@ public class PrometheusHttpListenerTests
             port = random.Next(2000, 5000);
             address = $"http://localhost:{port}/";
 
-            provider = Sdk.CreateMeterProviderBuilder()
-                .AddMeter(meter.Name)
-                .AddPrometheusHttpListener(options =>
-                {
-                    options.OpenMetricsEnabled = useOpenMetrics;
-                    options.UriPrefixes = new string[] { address };
-                })
-                .Build();
+            try
+            {
+                provider = Sdk.CreateMeterProviderBuilder()
+                    .AddMeter(meter.Name)
+                    .AddPrometheusHttpListener(options => options.UriPrefixes = new string[] { address })
+                    .Build();
+
+                break;
+            }
+            catch
+            {
+                // ignored
+            }
+        }
+
+        if (provider == null)
+        {
+            throw new InvalidOperationException("HttpListener could not be started");
         }
 
         var tags = new KeyValuePair<string, object>[]
@@ -137,9 +155,9 @@ public class PrometheusHttpListenerTests
 
         using HttpClient client = new HttpClient();
 
-        if (useOpenMetrics)
+        if (!string.IsNullOrEmpty(acceptHeader))
         {
-            client.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/openmetrics-text"));
+            client.DefaultRequestHeaders.Add("Accept", acceptHeader);
         }
 
         using var response = await client.GetAsync($"{address}metrics");
@@ -149,7 +167,7 @@ public class PrometheusHttpListenerTests
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
             Assert.True(response.Content.Headers.Contains("Last-Modified"));
 
-            if (useOpenMetrics)
+            if (requestOpenMetrics)
             {
                 Assert.Equal("application/openmetrics-text; version=1.0.0; charset=utf-8", response.Content.Headers.ContentType.ToString());
             }
@@ -160,12 +178,9 @@ public class PrometheusHttpListenerTests
 
             var content = await response.Content.ReadAsStringAsync();
 
-            var expected = useOpenMetrics
-                ? "# TYPE otel_scope_info info\n"
-                  + "# HELP otel_scope_info Scope metadata\n"
-                  + $"otel_scope_info{{otel_scope_name='{this.meterName}'}} 1\n"
-                  + "# TYPE counter_double_total counter\n"
-                  + $"counter_double_total{{otel_scope_name='{this.meterName}',key1='value1',key2='value2'}} 101.17 \\d+\\.\\d{{3}}\n"
+            var expected = requestOpenMetrics
+                ? "# TYPE counter_double_total counter\n"
+                  + "counter_double_total{key1='value1',key2='value2'} 101.17 \\d+\\.\\d{3}\n"
                   + "# EOF\n"
                 : "# TYPE counter_double_total counter\n"
                   + "counter_double_total{key1='value1',key2='value2'} 101.17 \\d+\n"
@@ -177,5 +192,7 @@ public class PrometheusHttpListenerTests
         {
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
         }
+
+        provider.Dispose();
     }
 }